refactor(templateLoader): use eval instead of require
The original template loader used require to load the chart
forms.  This caused the modules implementing forms to be
cached and therefore prevented any future reloading.
What's more the exporting the results of template
parsing resulted in those definitions being sticky as well.

The replaces the require with an eval based mechanism
and removes all exporting of template form+summary definitions

Signed-off-by: Kevin O'Donnell <kevin.odonnell@btp.works>